emotion_ai_project/requirements.txt
pyyaml>=5.4
requests
google-generativeai
